Started working on collision. Added option in renderer to render bounding boxes, as well as a renderer for rendering these lines (and lines in general).
Also added some basic set theory maths for mapping one set to another. The plan for this
is for coupling together components in entities, such that they can interact.

This is still in a very early and experimentive stage, but the idea is for components
to be able to create "events" through either input or interaction between entities (like collisions),
and each type of entity will have its own event-to-"desire"-map functions, where desires are things the entity "desires" to do, the desires dont necessarily need to be fulfilled.
An example of this is that when a key is pressed (for example 'W'), the input component creates a "'W' pressed" event, which the entity
can convert to a desire through the event-to-desire-map for that specific entity. In this example it may be converted to a "Move 10 pixels upwards" desire,
which the collision and physics components again will interact with.